http/mapping_rule: add an initial implementation of HTTP mapping rules
This introduces (default) optional features useful for proxies:

- rest-mappings
- rest-mappings-serde

Compatibility: while these do not enforce the "std" feature, they
currently end up requiring it effectively because of the regular
expression engine requirements.

This implements the matching of HTTP (REST) mapping rules via inspection
of the HTTP method, path and query string.

The current implementation is just an approximation of what Apicast is
currently doing. There are a few reasons for this, not the least that
the documentation underspecifies this functionality and that Apicast
itself just passes down unchecked contents to its PCRE-based regular
expression engine.

In particular, this specification allows users to specify:

- Method: the HTTP verb literal string or the string "any"
  (case-insensitive).
- Pattern: the path and the query string pattern to match against the
  actual path and query string of requests.

Methods must match the precise HTTP verb string, case-insensitive, or
use the special method "any", which matches all methods. Note that when
you set up a rule, you can use any other string as well, allowing for
new verbs.

Patterns are a bit of a mess, though. The general rule is that you can
use placeholders for values in paths matching a superset of the
alphanumeric characters, and use the same placeholders in query string
parameter values (not keys) that match just word characters and
underscore and dash. You do so by indicating them inside curly braces
like "/path/{id}?key={val}" - at least one character is required in
between the braces, though, and the matches also require at least one
character.

In the case of the path, streams of repeated forward slashes will be
converted to a single slash, and the $ dollar sign at the end of the
path part of the pattern is guaranteed to mean an exact match (ie.
"pear$" will match "pear" but not "pears"). Whatever your pattern it
will always match at the beginning of your path, so "/hello" will match
"/hello/world" but it won't match "/world/hello".

In the case of the query string, parameter keys and values will match
any order in which they appear, so "/path$?a=1&b={_}" will match
"/path?a=1&b=2" and it will also match "/path?b=3&a=1". Specified
patterns are requirements, so the presence of other unspecified
parameters does not make a difference.

However while with path patterns there's some sense into implicitly
assuming that the literal pattern should be matching, with query string
parameter values there is no such information, and in fact Apicast seems
to only look for whether a pattern exists or not to consider any value a
match.

Just like this point above, many things are not well defined or
questionable:

- What if the $ sign at the end of a pattern is preceded by a \?
- What if the $ sign appears elsewhere in the path?
- What if regular expression special characters are used outside the
  curly braces placeholders, like `[` or `(`?
- Do query string patterns also match from the beginning? Do they also
  support using $ at the end? At the end of the parameter name or value?
  Do both names and values have implicit matches to their respective
  beginnings? Do placeholders in values match assuming the whole value
  is variable or can placeholders work like path patterns with some
  literal parts, ie. /numbers-ending-with-1$?n={num}1
- Do placeholders always need at least one character? (ie. see previous
  point where n=1 would not match)
- Can you use placeholders in query string parameter keys?

There are quite a lot of places where this underspecification can bite a
user and the implementor.

The documentation also states that you can use placeholders recursively,
which both Apicast and this implementation don't support (at least not
correctly), but you'd be hardpressed to find value in that. The
documenttion also states other quite suspect things like Apicast looking
at POST/DELETE/etc bodies to match query string parameters (which this
implementation completely does not even pretend to do).

So because there is so much not specified, underspecified or just plain
wrong in the documentation, there are quite a few issues with the
current mapping rules implementations, including this one.

Apicast just looks for placeholders in paths and in query string
parameter values. With path patterns it uses a specific regular
expression to capture path segments (or parts), and with query string
parameter values it just looks at whether the value has a placeholder,
in which case everything matches directly, or if not it compares the
value literally for a match.

The thing is that whenever the pattern has been modified to contain a
regular expression, Apicast will then submit the whole string down to
its PCRE engine for evaluation. This way they support the above
definition, but they do so at the expense of leaving the interpretation
of all characters outside the placeholders to the regex engine. So a $
character outside a placeholder would only match the end of the string,
but if it is escaped with a backslash then it just means a literal $
sign. So because Apicast just passes down everything unescaped, lots of
additional behaviour is now trapped behind the interpretation of the
exact engine used.

So our implementation behaves the same by replacing placeholders and
submitting everything else down to our engine (which should be mostly
PCRE-compatible). Doing so this way feels wrong and is not specified,
that is, no one told the user that `/path[0-9]/id` would match
`path5/id`, for example, but instead they'd think it matches the literal
path including the brackets.

In the case of query string parameters this implementation also allows
placeholders in parameter keys but it escapes absolutely everything not
within the placeholders, which breaks the case of ending these patterns
with a $ character. On the other hand, using special characters outside
the placeholders won't have special meaning. It also allows for
placeholders in value positions, but they are fully evaluated to match
so that you can match on values that have literal parts, ie. `n=9{_}9`
for values that start with a 9 and have at least another one (but note
it does not mean it will end with a 9 necessarily).

Of course, the above is not coherent with how path patterns are being
matched. But there is little coherence overall, so let's change it
whenever the dust settles around this topic.

All in all, this needs discussion, agreement and docs fixing.